Fix ESBJAVA-5138
The trasnaction id used for the communication between SAP and the
SAPTransaportSender to send IDOCs can be of intereset when it comes to
responding to the client back in a Proxy. It could be used to track
what transactions were successful and what failed. The improvement was
to set it as a transport header to the synapse context so that it can
be refererded to, in the synapse sequence.
Addresses the issue at: https://wso2.org/jira/browse/ESBJAVA-5138